backend/resource_server/requirements.txt
deta
fastapi
uvicorn[standard]
email-validator
python-jose[cryptography]
websockets
gunicorn
flower